requirements.txt
# PyTorch Style Transfer Training
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# TensorFlow Style Transfer Training
tensorflow>=2.13.0

# Core ML Conversion
coremltools>=7.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0

# Configuration
pyyaml>=6.0
jsonschema>=4.17.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.22.0
nbformat>=5.8.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.2.0

# Optional: ONNX export
onnx>=1.14.0
onnxruntime>=1.15.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0
